도메인 정리

+++ src/main/java/com/kakao/moneyspread/domain/chatroom/ChatRoomRepository.java
@@ -2,6 +2,6 @@ package com.kakao.moneyspread.domain.chatroom;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 
 }
