+++ src/main/java/com/kakao/moneyspread/domain/spread/SpreadRepository.java
@@ -2,5 +2,5 @@ package com.kakao.moneyspread.domain.spread;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpreadRepository extends JpaRepository<Spread, Integer> {
+public interface SpreadRepository extends JpaRepository<Spread, Long> {
 }
